+++ src/components/molecules/DropdownFiltered/Options.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { colors } from '../../../constants/colors';
+import { colors } from '../../../constants';
 
 export interface OptionsListProps {
   optionsListMaxHeight?: string;
