+++ src/components/molecules/DropdownFiltered/SearchInput.ts
@@ -1,5 +1,5 @@
 import styled, { css } from 'styled-components';
-import { colors } from '../../../constants/colors';
+import { colors } from '../../../constants';
 import InputText from '../../atoms/InputText/InputText';
 import { InputProps } from '../../types';
