+++ src/components/molecules/Modal/ModalStyles/ModalStyles.tsx
@@ -1,6 +1,5 @@
 import { createGlobalStyle } from 'styled-components';
-
-import { colors } from '../../../../constants/colors';
+import { colors } from '../../../../constants';
 
 export interface ModalStylesProps {
   /**
