+++ src/components/molecules/ZopaFooter/styles/Footer.tsx
@@ -1,7 +1,5 @@
 import styled from 'styled-components';
-import { spacing } from '../../../../constants/spacing';
-import grid from '../../../../constants/grid';
-import { colors } from '../../../../constants/colors';
+import { colors, grid, spacing } from '../../../../constants';
 
 export const Footer = styled.footer`
   background-color: ${colors.white};
