+++ src/components/molecules/ZopaFooter/styles/LegalBlock.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import FlexCol from '../../../layout/FlexCol/FlexCol';
-import grid from '../../../../constants/grid';
+import { grid } from '../../../../constants';
 
 export const LegalBlock = styled(FlexCol).attrs({ xs: 12, l: 5 })`
   order: 3;
